Define a second command
