+++ lib/commands/embed_command.rb
@@ -16,8 +16,30 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-require 'ostruct'
-require 'fileutils'
+require('jekyll')
+require_relative('fragmentation')
+require_relative('embedding')
+require_relative('configuration')
 
-require_relative('commands/embed_command.rb')
-require_relative('commands/check_command.rb')
+# Usage example:
+#   bundle exec jekyll embedCodeSamples
+#
+
+module Jekyll::Commands
+
+  # Command which updates code embeddings in the documentation files.
+  class EmbedCodeSamples < Jekyll::Command
+
+    def self.init_with_program(prog)
+      prog.command('embedCodeSamples') do |c|
+        c.description 'Embeds sample code into doc files.'
+        c.action { |_, __| process(Configuration.from_file) }
+      end
+    end
+
+    def self.process(configuration)
+      Fragmentation.write_fragment_files configuration
+      EmbeddingProcessor.embed_all configuration
+    end
+  end
+end
